requirements.txt
pyautogui
opencv-python
pillow
openai
python-dotenv
